+++ lua/terrortown/lang/fr.lua
@@ -1860,3 +1860,8 @@ L.karma_unknown_tooltip = "Inconnu"
 --L.sb_rank_tooltip_team = "Team"
 
 --L.tbut_adminarea = "ADMIN AREA:"
+
+-- 2023-03-09
+--L.help_enforce_playercolor = [[
+--This sets all players to the same color on (re)spawn to prevent any distinction between playermodels and therefore the respective players.
+--If disabled this should allow colors from custom colorable playermodels to stick. Be aware that this might break (with) some appearance changing equipment e.g. make a disguise pretty obvious.]]