+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ### Added
 - Added a new font in default_skin.lua to fit the localization (by @Satton2)
+- Added `ttt_enforce_playercolor (default: 1)` convar
+  - can be turned off to stop ttt2 from overriding playermodel colors on (re)spawn
 
 ### Changed